S04_US20 - ADMIN Services, secured delete button

+++ src/Controller/AdminController.php
@@ -36,9 +36,11 @@ class AdminController extends AbstractController
     }
     public function deleteService(int $id): void
     {
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $servicesManager = new ServicesManager();
             $servicesManager->delete($id);
             header('Location: /Admin/index');
+        }
     }
     public function addService(): string
     {
